[poet.blessed.install] replaces [poet.blessed].
This is the first step in refactoring the [poet]
command suite to simplify the code in blessed
clients. Right now the clients have to source
[poet.blesser] and then call a function defined
in [poet.blesser].

We wish for clients to have to do only one thing,
not too. In preparation, we want to better name
the [poet.*] files to be more memorable to their
function. In addition, all poet functions that
have to do with blessing will have names of the
form [poet.blessed.*].